ref: Move serializers inside testing namespace
