+++ tests/integration/Utils/DER/DerSignatureSerializer.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Mdanter\Ecc\Serializer\Signature;
+namespace Mdanter\Ecc\Integration\Utils\DER;
 
 use Mdanter\Ecc\Crypto\Signature\Signature;
 use Mdanter\Ecc\Crypto\Signature\SignatureInterface;
