+++ tests/integration/Utils/fixtures/openssl-secp256r1.1.pub.pem
